Move getTitle()_and getDomain() to services
